requirements.txt
matplotlib==3.5.2
numpy==1.23.1
Pillow==9.2.0
tensorflow==2.9.1
git+https://github.com/qubvel/segmentation_models.git